Add Dialer.StartTLSPolicy for securing STARTTLS
It is possible to use a man-in-the-middle attack to strip STARTTLS
security from an SMTP connection. Because all modern SMTP servers
support STARTTLS, we recommend that `Dialer.StartTLSPolicy` be set to
`mail.MandatoryStartTLS`.

For backwards compatibility, this library defaults to
`mail.OpportunisticStartTLS`, which means that the client will
fallback to a cleartext connection if the SMTP server does not support
the STARTTLS extension.

This has no effect on SMTPS connections where TLS wraps the connection
itself.

Fixes #25

[Ivy:
 - squash Dialer option and constant renaming to StartTLSPolicy
 - fix changelog entry to reflect rename
 - squash Dial returns StartTLSUnsupportedError on failure]